+++ src/dataclocklib/charts.py
@@ -1,24 +1,5 @@
 """Data clock module for chart creation.
 
-Author: Andrew Ridyard.
-
-License: GNU General Public License v3 or later.
-
-Copyright (C): 2025.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
 Functions:
     dataclock: Create a data clock chart from a pandas DataFrame.
     line_chart: Create a line chart from a pandas DataFrame.
@@ -27,6 +8,9 @@ Constants:
     VALID_AGGREGATIONS: Tuple of valid aggregation function names.
     VALID_CMAPS: Tuple of valid colour map names.
     VALID_MODES: Tuple of valid chart modes.
+
+License:
+    SPDX-License-Identifier: GPL-3.0-or-later
 """
 
 from __future__ import annotations